Fixed sysvars_wsrep failure on 32bit
Made variable ulonglong instead of unstable ulong. There's no
value in using ulong anyway.

+++ sql/wsrep_trans_observer.h
@@ -328,7 +328,7 @@ static inline void wsrep_open(THD* thd)
     {
       thd->wsrep_cs().enable_streaming(
         wsrep_fragment_unit(thd->variables.wsrep_trx_fragment_unit),
-        thd->variables.wsrep_trx_fragment_size);
+        size_t(thd->variables.wsrep_trx_fragment_size));
     }
   }
   DBUG_VOID_RETURN;
